+++ Sources/Puredux/Store/StoreOf.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 
-
 @propertyWrapper
 public struct StoreOf<T> {
     private let keyPath: WritableKeyPath<Injected, T>
@@ -23,54 +22,4 @@ public struct StoreOf<T> {
     public init<State, Action>(_ keyPath: WritableKeyPath<Injected, T>) where T == StateStore<State, Action>? {
         self.keyPath = keyPath
     }
-}
-
-/** Provides access to injected dependencies. */
-public struct Injected {
-    
-    /** This is only used as an accessor to the computed properties within extensions of `InjectedValues`. */
-    private static var current = Injected()
-    
-    /** A static subscript for updating the `currentValue` of `InjectionKey` instances. */
-    static subscript<K>(key: K.Type) -> K.Value where K : InjectionKey {
-        get { key.currentValue }
-        set { key.currentValue = newValue }
-    }
-    
-    /** A static subscript accessor for updating and references dependencies directly. */
-    static subscript<T>(_ keyPath: WritableKeyPath<Injected, T>) -> T {
-        get { current[keyPath: keyPath] }
-        set { current[keyPath: keyPath] = newValue }
-    }
-}
-
- 
-public extension StoreOf {
-    func with<U, State, Action>(
-        _ state: U,
-        reducer: @escaping Reducer<U, Action>) -> StateStore<(State, U), Action> where T == StateStore<State, Action> {
-        
-        wrappedValue.stateStore(
-            state,
-            reducer: reducer
-        )
-    }
-    
-    func with<U, State, Action>(
-        _ state: U,
-        reducer: @escaping Reducer<U, Action>) -> StateStore<(State, U), Action>? where T == StateStore<State, Action>? {
-        
-        wrappedValue?.stateStore(
-            state,
-            reducer: reducer
-        )
-    }
-}
-
-public protocol InjectionKey {
-    /** The associated type representing the type of the dependency injection key's value. */
-    associatedtype Value
-
-    /** The default value for the dependency injection key. */
-    static var currentValue: Self.Value { get set }
 }
